fix(dashboard-perf): SQLX-Decode — AVG(peak/hours) als float8 statt 0
Im Wochentags-Chart (avgPeak/avgHours) und im Stunden-Heatmap (avgPeak)
las Rust AVG(s.peak_viewers) bzw. AVG(s.duration_seconds/3600.0) per
try_get::<f64>. peak_viewers ist INTEGER → AVG() liefert numeric,
duration/3600.0 ebenfalls numeric; sqlx kann numeric nicht als f64
dekodieren → unwrap_or(0.0) → die Felder waren konstant 0. Python liefert
dort echte Werte (float()).

Fix: AVG(...)::float8 (analog zum bereits korrekten SUM(...)::float8 für
airtime). avg_viewers bleibt ungecastet — die Spalte ist DOUBLE
PRECISION, AVG() liefert float8, der Read stimmt. retention_10m (REAL)
ist ebenfalls nicht betroffen (AVG(real) → double precision).

Latent bis v2-Caddy-Flip (v2 serviert noch Python 8765). cargo check
grün, Release-Build + Restart sauber.

+++ rust/crates/tb-dashboard-api/src/handlers/performance.rs
@@ -158,9 +158,9 @@ pub async fn weekly_stats_handler(
         SELECT
             EXTRACT(DOW FROM s.started_at)::integer AS weekday,
             COUNT(*) AS stream_count,
-            AVG(s.duration_seconds / 3600.0) AS avg_hours,
+            AVG(s.duration_seconds / 3600.0)::float8 AS avg_hours,
             AVG(s.avg_viewers) AS avg_viewers,
-            AVG(s.peak_viewers) AS avg_peak,
+            AVG(s.peak_viewers)::float8 AS avg_peak,
             SUM(CASE WHEN s.follower_delta IS NOT NULL
                      AND NOT (s.followers_end = 0 AND s.followers_start > 0)
                      THEN s.follower_delta ELSE 0 END) AS total_followers
@@ -224,7 +224,7 @@ pub async fn hourly_heatmap_handler(
             EXTRACT(HOUR FROM s.started_at)::integer AS hour,
             COUNT(*) AS stream_count,
             AVG(s.avg_viewers) AS avg_viewers,
-            AVG(s.peak_viewers) AS avg_peak
+            AVG(s.peak_viewers)::float8 AS avg_peak
         FROM twitch_stream_sessions s
         WHERE s.started_at >= $1
           AND s.ended_at IS NOT NULL
